review: move test to te existing NAD viewer test file
